Fixing merge problems
