--- client/src/utils/API.js
@@ -1,4 +1,0 @@
-export const searchBored = (query) => {
-  return fetch(`http://www.boredapi.com/api/activity/?key=5881028`);
-};
-
