+++ package.json
@@ -1,48 +1,36 @@
 {
   "name": "chairman-of-the-bored",
-  "version": "0.1.0",
-  "private": true,
+  "version": "1.0.0",
+  "description": "## User Story",
+  "main": "server/server.js",
+  "scripts": {
+    "start": "node server/server.js",
+    "develop": "concurrently \"cd server && npm run watch\" \"cd client && npm start\"",
+    "build": "cd client && npm run build",
+    "install": "cd server && npm i && cd ../client && npm i"
+  },
+  "repository": {
+    "type": "git",
+    "url": "git+https://github.com/Ginger421/chairman-of-the-bored.git"
+  },
+  "keywords": [],
+  "author": "",
+  "license": "ISC",
+  "bugs": {
+    "url": "https://github.com/Ginger421/chairman-of-the-bored/issues"
+  },
+  "homepage": "https://github.com/Ginger421/chairman-of-the-bored#readme",
   "dependencies": {
-    "@testing-library/jest-dom": "^5.16.5",
-    "@testing-library/react": "^13.4.0",
-    "@testing-library/user-event": "^13.5.0",
+    "@apollo/client": "^3.7.0",
     "apollo-server-express": "^3.10.2",
-    "autoprefixer": "^10.4.12",
+    "colors": "^1.4.0",
     "dotenv": "^16.0.3",
     "express": "^4.18.1",
     "graphql": "^16.6.0",
-    "mongodb": "^4.10.0",
-    "mongoose": "^6.6.3",
-    "postcss": "^8.4.17",
-    "react": "^18.2.0",
-    "react-dom": "^18.2.0",
-    "react-router-dom": "^6.4.1",
-    "react-scripts": "5.0.1",
-    "tailwindcss": "^3.1.8",
-    "web-vitals": "^2.1.4"
+    "jsonwebtoken": "^8.5.1",
+    "mongoose": "^6.6.4"
   },
-  "scripts": {
-    "start": "react-scripts start",
-    "build": "react-scripts build",
-    "test": "react-scripts test",
-    "eject": "react-scripts eject"
-  },
-  "eslintConfig": {
-    "extends": [
-      "react-app",
-      "react-app/jest"
-    ]
-  },
-  "browserslist": {
-    "production": [
-      ">0.2%",
-      "not dead",
-      "not op_mini all"
-    ],
-    "development": [
-      "last 1 chrome version",
-      "last 1 firefox version",
-      "last 1 safari version"
-    ]
+  "devDependencies": {
+    "concurrently": "^7.4.0"
   }
 }
